Organización tipos de datos de los id's

+++ ActionFactory/src/main/java/com/IntegrativeProject/ActionFactory/service/RoleService.java
@@ -25,7 +25,7 @@ public class RoleService {
         return  this.roleRepository.findAll();
     }
 
-    public Optional<Role> findById(int id) {
+    public Optional<Role> findById(Long id) {
         return roleRepository.findById(id);
     }
 }
